+++ resources/module_codes/en/extract.sh
@@ -4,7 +4,11 @@ function extract_single_img {
   local base_name="${single_file_name%.*}"
   fs_type=$(recognize_file_type "$single_file")
   start=$(date +%s%N)
-  rm -rf "$WORK_DIR/$current_workspace/Extracted-files/$base_name"
+  if [[ "$fs_type" == "ext" || "$fs_type" == "erofs" || "$fs_type" == "f2fs" || \
+        "$fs_type" == "boot" || "$fs_type" == "dtbo" || "$fs_type" == "recovery" || \
+        "$fs_type" == "vbmeta" || "$fs_type" == "vendor_boot" ]]; then
+    rm -rf "$WORK_DIR/$current_workspace/Extracted-files/$base_name"
+  fi
   case "$fs_type" in
     sparse)
       echo "Converting sparse partition file ${single_file_name}, please wait..."
@@ -12,16 +16,11 @@ function extract_single_img {
       rm -rf "$single_file"
       mv "$WORK_DIR/$current_workspace/${base_name}_converted.img" "$WORK_DIR/$current_workspace/${base_name}.img"
       single_file="$WORK_DIR/$current_workspace/${base_name}.img"
-      fs_type=$(recognize_file_type "$single_file")
-      if [ "$fs_type" == "super" ]; then
-        echo "Extracting non-sparse SUPER partition file ${single_file_name}, please wait..."
-        "$TOOL_DIR/7z" e -bb1 -aoa "$single_file" -o"$WORK_DIR/$current_workspace"
-        rm "$single_file"
-        mkdir -p "$WORK_DIR/$current_workspace/Extracted-files/super"
-      fi
+      extract_single_img "$single_file"
+      return
       ;;
     super)
-      echo "Extracting non-sparse SUPER partition file ${single_file_name}, please wait..."
+      echo "Extracting SUPER partition file ${single_file_name}, please wait..."
       "$TOOL_DIR/7z" e -bb1 -aoa "$single_file" -o"$WORK_DIR/$current_workspace"
       rm "$single_file"
       mkdir -p "$WORK_DIR/$current_workspace/Extracted-files/super"
@@ -68,6 +67,8 @@ function extract_single_img {
       mv -f "$file" "${file%_a.img}.img"
     elif [[ $base_name == *_a.ext ]]; then
       mv -f "$file" "${file%_a.ext}.img"
+    elif [[ $base_name == *.ext ]]; then
+      mv -f "$file" "${file%.ext}.img"
     fi
   done
   echo "${single_file_name} extraction completed"
